refactor(agents): preserve stop-reason recovery seam
